Added backend changes until now
