+++ findit-backend/Handlers/ProductHandler/IProductHandler.cs
@@ -1,0 +1,10 @@
+﻿namespace findit_backend.Handlers.ProductHandler
+{
+    public interface IProductHandler
+    {
+        bool checkProductExistance(string productId);
+        bool isPerishable(string productId);
+        int getProductStock(string productId);
+
+    }
+}
