+++ findit-backend/Helpers/ProductValidationHelper/IProductValidationHelper.cs
@@ -1,0 +1,7 @@
+﻿namespace findit_backend.Helpers.ProductValidationHelper
+{
+    public interface IProductValidationHelper
+    {
+        string ValidateProductAddition(string productId, int quantity);
+    }
+}
