+++ findit-backend/Managers/AddToCartManager/IAddToCartManager.cs
@@ -1,0 +1,7 @@
+﻿namespace findit_backend.Managers.AddToCartManager
+{
+    public interface IAddToCartManager
+    {
+        string AddProductToCart(string productId, int quantity);
+    }
+}
