+++ findit-backend/appsettings.json
@@ -7,7 +7,7 @@
     },
     "AllowedHosts": "*",
   "ConnectionStrings": {
-    "FindItDatabase": "Data Source=MSI\\SQLEXPRESS;Initial Catalog=FindIt;Integrated Security=True;Encrypt=False"
+    "FindItDatabase": "Data Source=FelipeMena\\SQLEXPRESS;Initial Catalog=FindIt;Integrated Security=True;Encrypt=False"
   },
     "Jwt": {
         "Issuer": "FindIt-Backend",
